pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "patchmatch"
version = "0.1"
description = "A lightweight patch-based image matching package using Siamese descriptors."
authors = [
    { name="Muhammad Rehan" }
]
requires-python = ">=3.8"
dependencies = [
    "opencv-python==4.11.0.86",
    "numpy==2.1.3",
    "tqdm==4.67.1",
    "torch==2.7.1",
    "pyyaml==6.0.2",
]

[tool.setuptools.packages.find]
where = ["patchmatch"]
